refactor(desktop): merge main/lib into main/shared
