+++ apps/desktop/src/main/ipc/waveform.ts
@@ -5,7 +5,7 @@ import type { WaveformPeaksResult } from '@shiranami/contracts';
 import { IPC_CHANNELS, WAVEFORM_PEAK_COUNT } from '@shiranami/contracts';
 import { handle } from './with-ipc-handler';
 import { logger } from '../logger';
-import { hashTrackKey, readCachedPeaks, writeCachedPeaks } from '../lib/waveform-cache';
+import { hashTrackKey, readCachedPeaks, writeCachedPeaks } from '../shared/waveform-cache';
 import { decodeWaveformPeaks, shutdownWaveformWorker } from '../waveform-host';
 import { waveformGetPeaksArgs } from './schemas/waveform';
