+++ apps/desktop/src/main/shared/README.md
@@ -1,0 +1,7 @@
+# `shared/`
+
+Cross-cutting primitives that carry rules or state: caches (waveform, folders),
+path/url safety, media-type maps, album-art hashing/derivation.
+
+Contrast with `utils/`, which holds stateless pure helpers (path resolution,
+throttle gates, spawn wrappers).
