+++ apps/desktop/src/main/shared/album-art-image.ts
