+++ apps/desktop/src/main/shared/waveform-cache.ts
